Removed the use of reflection
Finally learnt something useful in cs class - abstraction. I mean, I had a rough idea on how to use it and what not but this clarified things for me. Also found out that there's such a thing as a functional interface that allows you to call methods by name (a string) if you put it in a map. BY BY REFLECTION lol. Anyway, i gotta go back to all my other plugins now that calls methods by strings and replace it with this